+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support deleting single database entries
 - Support generating stats of reports without submitting them to the
   database first
+- Support HTML format in `analytics` subcommand
 
 ### Changed